[35] Separated frontend run details
The experiment run details component has been separated for easier maintainability.
